+++ CHANGELOG.md
@@ -8,7 +8,7 @@
 
 ### Improvements
 
-* [#]() versiondb NewKVStore accept string as store name.
+* [#1684](https://github.com/crypto-org-chain/cronos/pull/1684) versiondb NewKVStore accept string as store name.
 
 *Nov 6, 2024*